requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.16.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
idna==2.7
macholib==1.11
oauth2client==4.1.3
pefile==2018.8.8
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyInstaller==3.4
python-crontab==2.3.4
python-dateutil==2.7.3
python-telegram-bot==11.0.0
pywin32-ctypes==0.2.0
requests==2.19.1
rsa==4.0
schedule==0.5.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
